+++ invoicer_project/static/js/request_utils.js
@@ -1,0 +1,57 @@
+import { host } from './utils_clients.js'
+
+export async function obtainNewAccessToken() {
+    let response;
+    const data = {refresh: window.localStorage.getItem('refreshToken')};
+    try {
+        response = await fetch(host + '/user/refresh/', {
+            method: "POST",
+            body: JSON.stringify(data)
+        });
+        const newToken = await response.json();
+        const accessToken = newToken['access'];
+        window.localStorage.setItem('accessToken', accessToken);
+    } catch (error) {
+        console.error(error);
+    }
+    return response.status === 200;
+}
+
+export async function obtainUserInitials() {
+    let responseCode;
+    const token = window.localStorage.getItem('accessToken');
+    if (token) {
+        const data = { "accessToken": token };
+        try {
+            const serverReply = await fetch(host + '/user/decode/', {
+                method: "POST",
+                body: JSON.stringify(data)
+            });
+            responseCode = serverReply.status;
+            const initials = await serverReply.json();
+            if (responseCode === 200) {
+                fillInitials(initials);
+            } else if (responseCode === 400) {
+                const obtainedNewTokens = await obtainNewAccessToken();
+                if (!obtainedNewTokens) {
+                    window.location.href = host + '/user/login/';
+                } else {
+                    await obtainUserInitials();
+                }
+            } else {
+                window.location.replace(host + '/user/login/');
+            }
+        } catch (error) {
+            console.error(error);
+        }
+    }
+    else {
+        window.location.replace(host + '/user/login/');
+    }
+}
+
+function fillInitials(userData) {
+    const userFirstName = userData["first_name"];
+    const userLastName = userData["last_name"];
+    document.getElementById("user_name").textContent = userFirstName + " " + userLastName;
+}
